option to handle django's set_unusable_password
Django provides the option to set an unusable password for a user. This PR adds the ability to handle those kinds of passwords and gives the option to either raise an exception or simply return false for an unusable password. The default value is to raise an exception to make it compatible with the current behavior.